feat: add Administration tab with Device Commands and Danger Zone
