+++ src/renderer/components/RadioPanel.tsx
@@ -71,10 +71,6 @@ interface Props {
   isConnected: boolean;
   telemetryDeviceUpdateInterval?: number | null;
   deviceFixedPosition?: boolean | null;
-  onReboot: (seconds: number) => Promise<void>;
-  onShutdown: (seconds: number) => Promise<void>;
-  onFactoryReset: () => Promise<void>;
-  onResetNodeDb: (preserveFavorites?: boolean) => Promise<void>;
   ourPosition?: OurPosition | null;
   onSendPositionToDevice?: (lat: number, lon: number, alt?: number) => Promise<void>;
   deviceOwner?: { longName: string; shortName: string; isLicensed: boolean } | null;
@@ -83,9 +79,6 @@ interface Props {
     shortName: string;
     isLicensed: boolean;
   }) => Promise<void>;
-  onRebootOta?: (delay: number) => Promise<void>;
-  onEnterDfu?: () => Promise<void>;
-  onFactoryResetConfig?: () => Promise<void>;
   capabilities?: ProtocolCapabilities;
   meshcoreChannels?: { index: number; name: string; secret: Uint8Array }[];
   onMeshcoreSetChannel?: (idx: number, name: string, secret: Uint8Array) => Promise<void>;
@@ -618,16 +611,6 @@ function WifiPasswordField({
   );
 }
 
-interface PendingAction {
-  name: string;
-  title: string;
-  message: string;
-  confirmLabel: string;
-  danger?: boolean;
-  showPreserveFavorites?: boolean;
-  action: () => Promise<void>;
-}
-
 export default function RadioPanel({
   configTarget,
   onSetConfig,
@@ -637,17 +620,10 @@ export default function RadioPanel({
   channelConfigs,
   isConnected,
   deviceFixedPosition,
-  onReboot,
-  onShutdown,
-  onFactoryReset,
-  onResetNodeDb,
   ourPosition,
   onSendPositionToDevice,
   deviceOwner,
   onSetOwner,
-  onRebootOta,
-  onEnterDfu,
-  onFactoryResetConfig,
   capabilities,
   meshcoreChannels,
   onMeshcoreSetChannel,
@@ -895,9 +871,6 @@ export default function RadioPanel({
   const [status, setStatus] = useState<string | null>(null);
   const [applyingSection, setApplyingSection] = useState<string | null>(null);
 
-  // ─── Device command confirmation ──────────────────────────────
-  const [pendingAction, setPendingAction] = useState<PendingAction | null>(null);
-  const [nodeDbPreserveFavorites, setNodeDbPreserveFavorites] = useState(false);
   const { addToast } = useToast();
   const { t } = useTranslation();
   const deviceRoleOptions = useMemo(
@@ -1000,10 +973,6 @@ export default function RadioPanel({
     }
   };
 
-  const executeWithConfirmation = useCallback((action: PendingAction) => {
-    setPendingAction(action);
-  }, []);
-
   const handleSendAdvert = async () => {
     if (!onSendAdvert) return;
     setAdvertLoading(true);
@@ -1037,23 +1006,6 @@ export default function RadioPanel({
       setSyncClockLoading(false);
     }
   };
-
-  const handleConfirm = useCallback(async () => {
-    if (!pendingAction) return;
-    setPendingAction(null);
-    try {
-      await pendingAction.action();
-      addToast(t('radioPanel.actionCompleted', { name: pendingAction.name }), 'success');
-    } catch (err) {
-      console.warn('[RadioPanel] pending action failed ' + errLikeToLogString(err));
-      addToast(
-        t('radioPanel.actionFailed', {
-          message: err instanceof Error ? err.message : 'Unknown error',
-        }),
-        'error',
-      );
-    }
-  }, [pendingAction, addToast, t]);
 
   const handleImportConfig = useCallback(() => {
     const input = document.createElement('input');
@@ -2381,175 +2333,6 @@ export default function RadioPanel({
             )}
           </div>
         </div>
-      )}
-
-      {/* Device Commands — keep at bottom of Radio panel, directly above Danger Zone; do not reorder */}
-      <div className="space-y-3">
-        <h3 className="text-sm font-medium text-orange-400">{t('radioPanel.deviceCommands')}</h3>
-        <div className="space-y-2 rounded-lg border border-orange-900 p-4">
-          <p className="text-xs text-orange-400/80">
-            {t('radioPanel.deviceCommandsImmediateWarning')}
-          </p>
-          <div className="grid grid-cols-1 gap-2 md:grid-cols-2 lg:grid-cols-3 xl:grid-cols-4 2xl:grid-cols-5">
-            <button
-              type="button"
-              onClick={() => {
-                executeWithConfirmation({
-                  name: t('radioPanel.enterDfuName'),
-                  title: t('radioPanel.enterDfuTitle'),
-                  message: t('radioPanel.enterDfuMessage'),
-                  confirmLabel: t('radioPanel.enterDfuConfirm'),
-                  action: () => onEnterDfu?.() ?? Promise.resolve(),
-                });
-              }}
-              disabled={!isConnected || !onEnterDfu}
-              className="rounded-lg border border-orange-800/60 bg-orange-900/30 px-4 py-3 text-sm font-medium text-orange-200 transition-colors hover:bg-orange-900/50 disabled:opacity-50"
-            >
-              {t('radioPanel.enterDfuButton')}
-            </button>
-
-            <button
-              type="button"
-              onClick={() => {
-                executeWithConfirmation({
-                  name: t('radioPanel.rebootName'),
-                  title: t('radioPanel.rebootTitle'),
-                  message: capabilities?.hasCompanionContactManagementConfig
-                    ? t('radioPanel.rebootMessageMeshcore')
-                    : t('radioPanel.rebootMessageMeshtastic'),
-                  confirmLabel: t('radioPanel.rebootConfirm'),
-                  action: () => onReboot(2),
-                });
-              }}
-              disabled={!isConnected}
-              className="rounded-lg border border-orange-800/60 bg-orange-900/30 px-4 py-3 text-sm font-medium text-orange-200 transition-colors hover:bg-orange-900/50 disabled:opacity-50"
-            >
-              {t('radioPanel.rebootButton')}
-            </button>
-
-            <button
-              type="button"
-              onClick={() => {
-                executeWithConfirmation({
-                  name: t('radioPanel.rebootOtaName'),
-                  title: t('radioPanel.rebootOtaTitle'),
-                  message: t('radioPanel.rebootOtaMessage'),
-                  confirmLabel: t('radioPanel.rebootOtaConfirm'),
-                  action: () => onRebootOta?.(10) ?? Promise.resolve(),
-                });
-              }}
-              disabled={!isConnected || !onRebootOta}
-              className="rounded-lg border border-orange-800/60 bg-orange-900/30 px-4 py-3 text-sm font-medium text-orange-200 transition-colors hover:bg-orange-900/50 disabled:opacity-50"
-            >
-              {t('radioPanel.rebootOtaButton')}
-            </button>
-
-            {capabilities?.hasNodeDbReset !== false && (
-              <button
-                type="button"
-                onClick={() => {
-                  setNodeDbPreserveFavorites(false);
-                  executeWithConfirmation({
-                    name: t('radioPanel.resetNodeDbName'),
-                    title: t('radioPanel.resetNodeDbTitle'),
-                    message: t('radioPanel.resetNodeDbMessage'),
-                    confirmLabel: t('radioPanel.resetNodeDbConfirm'),
-                    showPreserveFavorites: configTarget?.mode === 'remote',
-                    action: () => onResetNodeDb(nodeDbPreserveFavorites),
-                  });
-                }}
-                disabled={!isConnected}
-                className="rounded-lg border border-orange-800/60 bg-orange-900/30 px-4 py-3 text-sm font-medium text-orange-200 transition-colors hover:bg-orange-900/50 disabled:opacity-50"
-              >
-                {t('radioPanel.resetNodeDbButton')}
-              </button>
-            )}
-
-            {capabilities?.hasShutdown !== false && (
-              <button
-                type="button"
-                onClick={() => {
-                  executeWithConfirmation({
-                    name: t('radioPanel.shutdownName'),
-                    title: t('radioPanel.shutdownTitle'),
-                    message: t('radioPanel.shutdownMessage'),
-                    confirmLabel: t('radioPanel.shutdownConfirm'),
-                    action: () => onShutdown(2),
-                  });
-                }}
-                disabled={!isConnected}
-                className="rounded-lg border border-orange-800/60 bg-orange-900/30 px-4 py-3 text-sm font-medium text-orange-200 transition-colors hover:bg-orange-900/50 disabled:opacity-50"
-              >
-                {t('radioPanel.shutdownButton')}
-              </button>
-            )}
-          </div>
-        </div>
-      </div>
-
-      {/* Danger Zone — keep at bottom of Radio panel after Device Commands; do not reorder */}
-      {capabilities?.hasFactoryReset !== false && (
-        <div className="space-y-3">
-          <h3 className="text-sm font-medium text-red-400">{t('radioPanel.dangerZone')}</h3>
-          <div className="space-y-2 rounded-lg border border-red-900 p-4">
-            <p className="text-xs text-red-400/80">{t('radioPanel.dangerZonePermanent')}</p>
-            <button
-              type="button"
-              onClick={() => {
-                executeWithConfirmation({
-                  name: t('radioPanel.factoryResetConfigName'),
-                  title: t('radioPanel.factoryResetConfigTitle'),
-                  message: t('radioPanel.factoryResetConfigMessage'),
-                  confirmLabel: t('radioPanel.factoryResetConfigConfirm'),
-                  danger: true,
-                  action: () => onFactoryResetConfig?.() ?? Promise.resolve(),
-                });
-              }}
-              disabled={!isConnected || !onFactoryResetConfig}
-              className="w-full rounded-lg border border-red-800/60 bg-red-900/40 px-4 py-3 text-sm font-medium text-red-300 transition-colors hover:bg-red-900/60 disabled:opacity-50"
-            >
-              {t('radioPanel.factoryResetConfigButton')}
-            </button>
-            <button
-              type="button"
-              onClick={() => {
-                executeWithConfirmation({
-                  name: t('radioPanel.factoryResetName'),
-                  title: t('radioPanel.factoryResetTitle'),
-                  message: t('radioPanel.factoryResetMessage'),
-                  confirmLabel: t('radioPanel.factoryResetConfirm'),
-                  danger: true,
-                  action: () => onFactoryReset(),
-                });
-              }}
-              disabled={!isConnected}
-              className="w-full rounded-lg border border-red-800 bg-red-900/50 px-4 py-3 text-sm font-medium text-red-300 transition-colors hover:bg-red-900/70 disabled:opacity-50"
-            >
-              {t('radioPanel.factoryResetButton')}
-            </button>
-          </div>
-        </div>
-      )}
-
-      {/* Confirmation Modal */}
-      {pendingAction && (
-        <ConfirmModal
-          title={pendingAction.title}
-          message={pendingAction.message}
-          confirmLabel={pendingAction.confirmLabel}
-          danger={pendingAction.danger}
-          preserveFavorites={
-            pendingAction.showPreserveFavorites ? nodeDbPreserveFavorites : undefined
-          }
-          onPreserveFavoritesChange={
-            pendingAction.showPreserveFavorites ? setNodeDbPreserveFavorites : undefined
-          }
-          onConfirm={handleConfirm}
-          onCancel={() => {
-            setPendingAction(null);
-            setNodeDbPreserveFavorites(false);
-          }}
-        />
       )}
     </div>
   );
