+++ src/renderer/locales/es/translation.json
@@ -2367,7 +2367,8 @@
     "repeaters": "Repetidoras",
     "rf": "RF",
     "graph": "Gráfico",
-    "rooms": "Salas"
+    "rooms": "Salas",
+    "admin": "Administración"
   },
   "takServerPanel": {
     "title": "Servidor TAK",
