+++ src/renderer/locales/fr/translation.json
@@ -2367,7 +2367,8 @@
     "repeaters": "Répéteurs",
     "rf": "RF",
     "graph": "Graphique",
-    "rooms": "Salles"
+    "rooms": "Salles",
+    "admin": "Administration"
   },
   "takServerPanel": {
     "title": "Serveur TAK",
