+++ src/renderer/locales/id/translation.json
@@ -2367,7 +2367,8 @@
     "repeaters": "Repeater",
     "rf": "RF",
     "graph": "Grafik",
-    "rooms": "Ruang"
+    "rooms": "Ruang",
+    "admin": "Administrasi"
   },
   "takServerPanel": {
     "title": "Server TAK",
