+++ src/renderer/locales/it/translation.json
@@ -2367,7 +2367,8 @@
     "repeaters": "Ripetitori",
     "rf": "RF",
     "graph": "Graph",
-    "rooms": "Camere"
+    "rooms": "Camere",
+    "admin": "Amministrazione"
   },
   "takServerPanel": {
     "title": "TAK Server",
