+++ src/renderer/locales/uk/translation.json
@@ -2367,7 +2367,8 @@
     "repeaters": "Ретранслятори",
     "rf": "RF",
     "graph": "Графік",
-    "rooms": "Кімнати"
+    "rooms": "Кімнати",
+    "admin": "Адміністрування"
   },
   "takServerPanel": {
     "title": "Сервер TAK",
